feat(frontend): relocate TodosList and AddTodo to pages as route components
