+++ frontend/src/AppRoutes.js
@@ -1,15 +1,15 @@
 import React from "react";
-import { Routes, Route } from "react-router-dom";
+import { Routes, Route, Navigate } from "react-router-dom";
 
-import AddTodo from "./components/AddTodo";
+import AddTodo from "./pages/AddTodo";
 import Login from "./pages/Login";
 import Signup from "./pages/Signup";
-import TodosList from "./components/TodosList";
+import TodosList from "./pages/TodosList";
 
 function AppRoutes({ token, user, login, signup }) {
   return (
     <Routes>
-      <Route path="/" element={<TodosList token={token} />} />
+      <Route path="/" element={<Navigate to="/todos" replace />} />
       <Route path="/todos" element={<TodosList token={token} />} />
       <Route path="/todos/create" element={<AddTodo token={token} />} />
       <Route path="/todos/:id" element={<AddTodo token={token} />} />
